Fixed integer type warning

+++ IADownloadManager/IADownloadManager.h
@@ -58,6 +58,6 @@ typedef void (^IACompletionBlock)(BOOL success, AFHTTPRequestOperation *operatio
 
 + (BOOL) isDownloadingItemWithURL:(NSURL*)url;
 + (void) stopDownloadingItemWithURL:(NSURL*)url;
-+ (int) listenerCountForUrl:(NSURL *)url;
++ (NSInteger) listenerCountForUrl:(NSURL *)url;
 
 @end
